requirements.txt
requests==2.31.0
python-dotenv==1.0.1
ccxt==4.2.14
pandas==2.2.0